Clarified fgsea rank ordering - Results unchanged
Original rank-ordering of the gene list in fgsea was arranged in increasing order (this was used for visualization purposes in an earlier instance of the code). For clarity, this update puts the gene list in the decreasing order to be in line with what users expect for GSEA. Results are unchanged because fgsea automatically ranks the stats argument.

Improved commenting.